+++ src/scenes/Scene.cpp
@@ -7,12 +7,8 @@
 #include <iostream>
 #include <glad/glad.h>
 
-Scene::Scene() {}
-
 void Scene::init(uint32 windowWidth, uint32 windowHeight) {
-  glViewport(0, 0, windowWidth, windowHeight);
-  this->windowWidth = windowWidth;
-  this->windowHeight = windowHeight;
+  framebufferSizeChange(windowWidth, windowHeight);
 }
 
 // Callback function for when user resizes our window
